Renamed oauth1a to two legged since it does not require user intervention
